Move new Price Calculator files into src/features/priceCalculator
