+++ apps/store/src/services/PriceCalculator/PriceCalculator.helpers.ts
@@ -10,8 +10,8 @@ import { SE_PET_CAT } from './data/SE_PET_CAT'
 import { SE_PET_DOG } from './data/SE_PET_DOG'
 import { SE_STUDENT_APARTMENT } from './data/SE_STUDENT_APARTMENT'
 import { SE_WIDGET_APARTMENT, SE_WIDGET_APARTMENT_NO_COMPARE } from './data/SE_WIDGET_APARTMENT'
-import type { InputField } from './Field.types'
 import { MIXED_BREED_OPTION_ID } from './Field.types'
+import type { InputField } from './Field.types'
 import type { Form, FormSection, JSONData, Template } from './PriceCalculator.types'
 
 const TEMPLATES: Record<string, Template | undefined> = {
